// src/api/endpoints.ts
export const enum Endpoints {
  GET_TABLE = "/get-table",
  ADD_ROW = "/add-row",
  DELETE_ROW = "/delete-row",
  EDIT_ROW = "/edit-row",
  SHOW_ROWS = "/show-rows",
  CREATE_TABLE = "/create-table",
  SAVE_BACKUP_TABLE = "/save-backup-table",
  CLEAR_TABLE = "/clear-table",
  DELETE_TABLE = "/delete-table",
  SET_BACKUP_TABLE = "/set-backup-table",
  DOES_TABLE_EXIST = "/exist",
}